Split ble & internet persmissions to prevent merging permission entries in Manifest
