--- permission/src/main/AndroidManifest.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2022, Nordic Semiconductor
-  ~ All rights reserved.
-  ~
-  ~ Redistribution and use in source and binary forms, with or without modification, are
-  ~ permitted provided that the following conditions are met:
-  ~
-  ~ 1. Redistributions of source code must retain the above copyright notice, this list of
-  ~ conditions and the following disclaimer.
-  ~
-  ~ 2. Redistributions in binary form must reproduce the above copyright notice, this list
-  ~ of conditions and the following disclaimer in the documentation and/or other materials
-  ~ provided with the distribution.
-  ~
-  ~ 3. Neither the name of the copyright holder nor the names of its contributors may be
-  ~ used to endorse or promote products derived from this software without specific prior
-  ~ written permission.
-  ~
-  ~ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-  ~ "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
-  ~ TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-  ~ PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-  ~ HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-  ~ SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-  ~ LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
-  ~ OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
-  ~ OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-  ~ NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
-  ~ EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-  -->
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
-
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-    <!--
-     Bluetooth permission is required on Android 4.3 - 11 in order to communicate with
-     Bluetooth LE devices.
-    -->
-    <uses-permission
-        android:name="android.permission.BLUETOOTH"
-        android:maxSdkVersion="30"/>
-    <!--
-     Bluetooth Admin permission is required on Android 4.3 - 11 in order to scan for
-     Bluetooth LE devices.
-    -->
-    <uses-permission
-        android:name="android.permission.BLUETOOTH_ADMIN"
-        android:maxSdkVersion="30"/>
-    <!--
-     Location permission is required from Android 6 to 11 to be able to scan for advertising
-     Bluetooth LE devices. Some BLE devices, called beacons, may be used to position the phone.
-     This is to ensure that the user agrees to do so.
-     This app does not use this location information in any way.
-     Since Android 10 ACCESS_FINE_LOCATION is required and for that, the COARSE one also needs
-     to be specified.
-    -->
-    <uses-permission
-        android:name="android.permission.ACCESS_COARSE_LOCATION"
-        android:maxSdkVersion="30"/>
-    <uses-permission
-        android:name="android.permission.ACCESS_FINE_LOCATION"
-        android:maxSdkVersion="30"/>
-    <!--
-     On Android 12 new set of Bluetooth permissions were added. BLUETOOTH_SCAN permission is
-     needed to scan for Bluetooth LE devices. The "never for location" flag is set, as we are not
-     interested in finding the user location, and we are OK with the fact, that beacons info
-     will be removed from the scan results.
-    -->
-    <uses-permission
-        android:name="android.permission.BLUETOOTH_SCAN"
-        android:usesPermissionFlags="neverForLocation"
-        tools:targetApi="s" />
-    <!--
-     This permission is required in order to connect to a Bluetooth LE device on Android 12
-     onwards.
-    -->
-    <uses-permission android:name="android.permission.BLUETOOTH_CONNECT" />
-
-    <uses-feature
-        android:name="android.hardware.bluetooth_le"
-        android:required="true"/>
-</manifest>
