--- permission/src/main/java/no/nordicsemi/android/common/permission/util/StandardPermissionState.kt
@@ -1,12 +1,0 @@
-package no.nordicsemi.android.common.permission.util
-
-enum class StandardPermissionNotAvailableReason {
-    NOT_AVAILABLE,
-    DISABLED,
-}
-
-sealed class StandardPermissionState {
-    object Available : StandardPermissionState()
-    data class NotAvailable(val reason: StandardPermissionNotAvailableReason) :
-        StandardPermissionState()
-}
