+++ permissions/internet/build.gradle.kts
@@ -30,35 +30,40 @@
  */
 
 plugins {
-    alias(libs.plugins.nordic.application.compose)
+    alias(libs.plugins.nordic.library.compose)
     alias(libs.plugins.nordic.hilt)
+    alias(libs.plugins.nordic.nexus)
 }
 
 group = "no.nordicsemi.android.common"
 
+nordicNexusPublishing {
+    POM_ARTIFACT_ID = "permissions-internet"
+    POM_NAME = "Nordic library for checking required internet permissions."
+
+    POM_DESCRIPTION = "Nordic Android Common Libraries"
+    POM_URL = "https://github.com/NordicPlayground/Android-Common-Libraries"
+    POM_SCM_URL = "https://github.com/NordicPlayground/Android-Common-Libraries"
+    POM_SCM_CONNECTION = "scm:git@github.com:NordicPlayground/Android-Common-Libraries.git"
+    POM_SCM_DEV_CONNECTION = "scm:git@github.com:NordicPlayground/Android-Common-Libraries.git"
+
+    POM_DEVELOPER_ID = "syzi"
+    POM_DEVELOPER_NAME = "Sylwester Zieliński"
+    POM_DEVELOPER_EMAIL = "sylwester.zielinski@nordicsemi.no"
+}
+
 android {
-    namespace = "no.nordicsemi.android.common.test"
+    namespace = "no.nordicsemi.android.common.permissions.ble"
 }
 
 dependencies {
+    implementation(project(":core"))
     implementation(project(":theme"))
-    implementation(project(":uilogger"))
-    implementation(project(":uiscanner"))
     implementation(project(":navigation"))
-    implementation(project(":permissions:nfc"))
-    implementation(project(":permissions:ble"))
-    implementation(project(":permissions:internet"))
 
     implementation(libs.androidx.compose.material.iconsExtended)
 
-    implementation(libs.androidx.activity.compose)
-
-    implementation(libs.nordic.blek.scanner)
-
+    implementation(libs.androidx.navigation.compose)
     implementation(libs.androidx.hilt.navigation.compose)
-
     implementation(libs.androidx.lifecycle.runtime.compose)
-
-    // debugImplementation because LeakCanary should only run in debug builds.
-    debugImplementation(libs.leakcanary)
 }
